Partial transparency check now uses alpha values of the color instead of checking if it is black.

+++ srcs/sector_wall_raycast.c
@@ -6,7 +6,7 @@
 /*   By: saaltone <saaltone@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 13:12:51 by saaltone          #+#    #+#             */
-/*   Updated: 2022/11/22 17:11:06 by saaltone         ###   ########.fr       */
+/*   Updated: 2022/11/22 18:37:44 by saaltone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -217,7 +217,7 @@ void	sector_walls_raycast(t_app *app, t_thread_data *thread, t_wall *wall,
 			|| wall->start_x > x || wall->end_x < x
 			|| app->occlusion_top[x] + app->occlusion_bottom[x] >= WIN_H
 			|| !raycast_hit(app, wall->line, &hit, x)
-			|| hit.distance > MAX_VIEW_DISTANCE)
+			|| hit.distance > (double)MAX_VIEW_DISTANCE)
 			continue ;
 		if (wall->is_portal)
 		{
